+++ src/backend/commands/common-admin-command.manager.ts
@@ -1,17 +1,22 @@
 import {CommonAdminCommand} from './common-admin.command';
+import * as Promise_serial from 'promise-serial';
 
 // TODO move to commons
+export interface CommonAdminPreparedCommandCommandConfigType {
+    parameters: {[key: string]: string}[];
+}
+
 export interface CommonAdminPreparedCommandConfigType {
-    parameters: {[key: string]: string};
+    commands: CommonAdminPreparedCommandCommandConfigType[];
     description: string;
 }
 
 // TODO move to commons
 export interface CommonAdminCommandConfigType {
     adminWritable: boolean;
-    availableCommands: {[key: string]: string[]};
+    availableCommands?: {[key: string]: string[]};
     preparedCommands: {[key: string]: CommonAdminPreparedCommandConfigType};
-    constantParameters: {[key: string]: string};
+    constantParameters?: {[key: string]: string};
 }
 
 // TODO move to commons
@@ -61,7 +66,7 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
             return res;
         }
 
-        for (const key in this.adminCommandConfig.availableCommands) {
+        for (const key in this.commands) {
             if (!this.adminCommandConfig.availableCommands.hasOwnProperty(key)) {
                 continue;
             }
@@ -86,11 +91,27 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
                 return me.initializePreparedCommand(initializedArgs);
             }
 
-            return me.initializeCommand(initializedArgs);
-        }).then(commandRequest => {
-            return me.validateCommandParameters(commandRequest);
-        }).then(commandRequest => {
-            return me.processCommandArgs(commandRequest);
+            return me.initializeCommand(initializedArgs).then(commandRequest => {
+                return Promise.resolve([commandRequest]);
+            });
+        }).then(commandRequests => {
+            const promises = [];
+            for (const commandRequest of commandRequests) {
+                promises.push(function() {
+                    return me.validateCommandParameters(commandRequest);
+                });
+            }
+
+            return Promise_serial(promises, {parallelize: 1});
+        }).then(commandRequests => {
+            const promises = [];
+            for (const commandRequest of commandRequests) {
+                promises.push(function() {
+                    return me.processCommandArgs(commandRequest);
+                });
+            }
+
+            return Promise_serial(promises, {parallelize: 1});
         });
     }
 
@@ -107,7 +128,7 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
         return Promise.resolve(initializedArgs);
     }
 
-    protected initializePreparedCommand(argv: {}): Promise<CommonAdminCommandsRequestType> {
+    protected initializePreparedCommand(argv: {}): Promise<CommonAdminCommandsRequestType[]> {
         const preparedCommandName = argv['preparedCommand'];
         if (preparedCommandName === undefined) {
             return Promise.reject('preparedCommand not defined');
@@ -131,27 +152,54 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
             return Promise.reject('preparedCommand not found');
         }
 
-        const initializedArgs = {};
-        for (const key in preparedCommand.parameters) {
-            if (!preparedCommand.parameters.hasOwnProperty(key)) {
-                continue;
+        const preparedCommandRequest: CommonAdminCommandsRequestType[] = [];
+        for (const command of preparedCommand.commands) {
+            const initializedArgs = {};
+            for (const key in command.parameters) {
+                if (!command.parameters.hasOwnProperty(key)) {
+                    continue;
+                }
+
+                initializedArgs[key] = command.parameters[key];
             }
 
-            initializedArgs[key] = preparedCommand.parameters[key];
+            const requestedCommand = initializedArgs['command'];
+            if (this.commands[requestedCommand] === undefined) {
+                return Promise.reject('command not defined');
+            }
+
+            preparedCommandRequest.push({
+                command: this.commands[requestedCommand],
+                parameters: initializedArgs
+            });
         }
 
-        const requestedCommand = initializedArgs['command'];
-        if (this.commands[requestedCommand] === undefined) {
-            return Promise.reject('command not defined');
-        }
-
-        return Promise.resolve({
-            command: this.commands[requestedCommand],
-            parameters: initializedArgs
-        });
+        return Promise.resolve(preparedCommandRequest)
     }
 
     protected initializeCommand(argv: {}): Promise<CommonAdminCommandsRequestType> {
+        const requestedCommand = argv['command'];
+        const requestedAction = argv['action'];
+        return this.validateCommandAndAction(requestedCommand, requestedAction).then(() => {
+            const command = this.commands[requestedCommand];
+            return Promise.resolve({
+                command: command,
+                parameters: argv
+            });
+        })
+    }
+
+    protected validateCommandParameters(commandRequest: CommonAdminCommandsRequestType): Promise<CommonAdminCommandsRequestType> {
+        return commandRequest.command.validateCommandParameters(commandRequest.parameters).then(() => {
+            return Promise.resolve(commandRequest);
+        });
+    }
+
+    protected processCommandArgs(commandRequest: CommonAdminCommandsRequestType): Promise<any> {
+        return commandRequest.command.process(commandRequest.parameters);
+    }
+
+    protected validateCommandAndAction(requestedCommand: string, requestedAction: string): Promise<boolean> {
         if (this.adminCommandConfig === undefined || this.adminCommandConfig.adminWritable !== true) {
             return Promise.reject('adminCommandConfig.adminWritable not active');
         }
@@ -166,7 +214,6 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
         }
 
         let availableCommandActions;
-        const requestedCommand = argv['command'];
         if (Object.keys(this.adminCommandConfig.availableCommands).length === 1
             && this.adminCommandConfig.availableCommands['*'] !== undefined) {
             availableCommandActions = this.adminCommandConfig.availableCommands['*'];
@@ -176,7 +223,6 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
             return Promise.reject('command not allowed');
         }
 
-        const requestedAction = argv['action'];
         if (requestedAction !== undefined
             && !(availableCommandActions.length === 1 && availableCommandActions[0] === '*')
             && !availableCommandActions.includes(requestedAction)) {
@@ -187,20 +233,11 @@ export abstract class CommonAdminCommandManager<A extends CommonAdminCommandConf
             return Promise.reject('command not defined');
         }
 
-        return Promise.resolve({
-            command: this.commands[requestedCommand],
-            parameters: argv
+        const command = this.commands[requestedCommand];
+        return command.validateCommandAction(requestedAction).then(() => {
+            return Promise.resolve(true);
         });
     }
 
-    protected validateCommandParameters(commandRequest: CommonAdminCommandsRequestType): Promise<CommonAdminCommandsRequestType> {
-        return commandRequest.command.validateCommandParameters(commandRequest.parameters).then(() => {
-            return Promise.resolve(commandRequest);
-        });
-    }
-
-    protected processCommandArgs(commandRequest: CommonAdminCommandsRequestType): Promise<any> {
-        return commandRequest.command.process(commandRequest.parameters);
-    }
 }